refactor(flow): lift evaluateWorkflowCondition + resolveWorkflowBindingValue + getPathValue + isRecord into @sentropic/flow/condition-eval (BR26 Lot 7 real reorg slice 1)
